+++ src/language/he.ts
@@ -1,0 +1,41 @@
+import assay from "./assay/he.json";
+import bank from "./bank/he.json";
+import credits from "./credits/he.json";
+import doctor from "./doctor/he.json";
+import end from "./end/he.json";
+import events from "./events/he.json";
+import health from "./health/he.json";
+import jail from "./jail/he.json";
+import mercantile from "./mercantile/he.json";
+import mines from "./mines/he.json";
+import newspaper from "./newspaper/he.json";
+import options from "./options/he.json";
+import other from "./other/he.json";
+import river from "./river/he.json";
+import saloon from "./saloon/he.json";
+import stable from "./stable/he.json";
+import startup from "./startup/he.json";
+import time from "./time/he.json";
+
+const lang = {
+  assay,
+  bank,
+  credits,
+  doctor,
+  end,
+  events,
+  health,
+  jail,
+  mercantile,
+  mines,
+  newspaper,
+  options,
+  other,
+  river,
+  saloon,
+  stable,
+  startup,
+  time,
+};
+
+export default lang;